requirements.txt
openai
python-dotenv
sentence-transformers
pytest